ENH: provide output for unknown types, scoped types, templated types
The main change here is that ArgClasses and ReturnClass are set for
all types, not just VTK_PARSE_VTK_OBJECT.  The main benefit of this
is that even if the type is VTK_PARSE_UNKNOWN, a string that gives
the class name will be available.  This will be true even for scoped
types, templated types, or scoped-and-templated types.  The only
types that are not output as such are function pointers.  They are
simply named "function".