ENH: Major update, merge all inherited methods into the XML
A new utility called vtkParseMerge allows the superclass
methods to be merged into a ClassInfo struct.  It also keeps
track of which class each method comes from.  In the XML output,
a new section called "Inheritance" lists the geneology in order
of method resolution.  Each method and variable has a <Class> tag
to identify what class defined the method or, in the case of
variables, defined a method that accesses the variable.