ENH: rename <Variable> to <Property> universally
There is an important distinction between "variables" and
"properties".  A property is identified by its access methods,
while a variable is "just there".  Properties are used in languages
such as Python, but the definition seems valid here, too.
Also, the name "variable" should be reserved for the actual
variables defined in the header files.