ENH: Move utility functions out of vtkParse.y.
The yacc file vtkParse.y was overly long, so many of the utility
functions that used to be defined in it are now in vtkParseData.c
and vtkParseString.c.  This also reduces the dependency on vtkParse.h,
and allows programs to use the vtkParse data structures without
needing to link to the parser itself.