Fix wrapper-parser rule for templated destructors.
The old function_name rule could not handle identifiers that started
with a tilde if the identifier was templated.  This fix replaces the
function_name rule with unqualified_id, and adds an optional tilde to
the template_id rule.  Because unqualified_id uses template_id, this
fixes the templated destructor problem.