Add special right angle bracket handling for C++11.
The C++11 standard allows a pair of right angles ">>" to be broken
into two right angle brackets in order to terminate a template
argument list.  In order to improve compatibility with C++11 code,
I've added the same flexibility to the wrapper's parser.  I have
checked the generated xml for VTK both before and after this change,
and the resulting xml was identical.