Add new C++11 string and char tokens to preprocessor.
This covers the u, U, and u8 prefixes plus the old L prefix, and
also reads the optional user-defined suffix, but only if the suffix
begins with an underscore.  When changing this, I noticed that the
code in vtkParse.y for writing constant expressions was not adding
a space between identifiers and literals.