Require leading underscore for user-defined string suffix.
The C++11 standard says that any sequence of ID characters immediately
following any literal are a user-defined literal suffix and are in fact
part of the literal.  For strings, this is a big problem, because some
headers have constructs like "string"MACRO where a string and a macro
ID are adjacent.  Both clang and gcc have taken the step of requiring
that all suffixes start with an underscore, to make it so that fewer
existing headers break the standard.  I've decided to do the same.