Change rules to disambiguate angle brackets from less-than.
Some constant_expressions could be parsed either as comparison or as
arguments in angle brackets.  This change takes advantage of the fact
that, in a constant_expression, angle brackets are generally followed
by "::", whereas greater-than is generally not.  It isn't perfect,
because "y > ::myclass::myconstant" will cause a syntax error now.