Remove "const" from parameter types
Whether a parameter is "const" does not change the API, it only
changes the constness within the body of the function.

Note that the constness must be removed at the right position,
for example "const char* x" remains the same, but "char* const x"
becomes "char* x".  That is, we want to remove the const that
cleaves to the parameter id, not the const that cleaves to the
type id.