Fix vtkParseProperties GetNth/SetNth detection
Methods like SetNthSomething(int) were detected as regular "SET"
methods instead of as "SET_NTH" methods because they had a single
parameter.  This was causing build errors when wrapping Slicer
because the property wrapping was trying to call a nonexistent
SetSomething() method.  The issue was fixed by rearranging the
order of the method type checks.